+++ lib/content-services/content-metadata/components/content-metadata-card/content-metadata-card.component.html
@@ -4,9 +4,6 @@
     </mat-card-content>
     <mat-card-footer class="adf-content-metadata-card-footer" fxLayout="row" fxLayoutAlign="space-between stretch">
         <div>
-            <button mat-icon-button>
-                <mat-icon>star_border</mat-icon>
-            </button>
             <button mat-icon-button (click)="toggleEdit()">
                 <mat-icon>mode_edit</mat-icon>
             </button>
