+++ lib/content-services/content-metadata/components/content-metadata-card/content-metadata-card.component.ts
