+++ lib/content-services/content-metadata/components/content-metadata/content-metadata.component.scss
