+++ lib/content-services/content-metadata/content-metadata.module.ts
@@ -21,8 +21,8 @@ import { NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 import { MaterialModule } from '../material.module';
 import { CardViewModule , FileSizePipe } from '@alfresco/adf-core';
-import { ContentMetadataComponent } from './content-metadata.component';
-import { ContentMetadataCardComponent } from './content-metadata-card.component';
+import { ContentMetadataComponent } from './components/content-metadata/content-metadata.component';
+import { ContentMetadataCardComponent } from './components/content-metadata-card/content-metadata-card.component';
 import { PropertyDescriptorsService } from './services/property-descriptors.service';
 import { PropertyDescriptorLoaderService } from './services/properties-loader.service';
 import { AspectWhiteListService } from './services/aspect-whitelist.service';
